Fixed eslint and webpack config
ESLint was throwing errors in any config.js files, because they lack
type information. Using "overrides" to define '.ts' rules solves this
problem by only applying rules to typescript files. May want to apply
*some* of the basic rules to '.js' files as well at some point.

Webpack has been failing to apply a 'title' to the index.html for some
time, but I just noticed. Pretty sure this worked at some point, but now
a line has to be added to the html.template for the parser to add in a
title.

    Note: It seems this pattern can be used to insert arbitrary keys
    into the html, which may be nice with separate dev and prod configs.