Add Delete Max Button
Apollo Cache is updated properly on delete.

AllMaxesQuery is updated to remove the deleted record, and the record itself is evicted from the cache. Eviction is not necessary in this case, since garbage collection would find things anyway, but it seems a good practice in cases where the record *does* exist in other queries.