Local Auth via Keycloak
Adds rudimentary local auth with Keycloak.

Graphql queries will get 403, and display 'Not Logged In' snackbar if user is not logged in.

For now, user can login by clicking the account menu icon in the navbar.

No meaningful authorization is present, yet.

Also changed HashRouter to BrowserRouter, as reading the react-router docs as part of this work identified HashRouter as an unrecommended option.